+++ CHANGELOG.md
@@ -19,6 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update research viewer role permissions. [#546](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/546)
 - Migrate and hide design plot fields. [#549](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/549)
 
+### Removed
+- Remove design as treatment fields. [#552](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/552)
+
 ## [2.16.2](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/milestone/34) 2023-09-06
 
 ### Added